# attached_assets-4/attached_assets/2/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clientum CRM/ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List deals
      parameters:
        - name: stage
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Company:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        contactCount:
          type: integer
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        industry:
          type: string
        size:
          type: string
        website:
          type: string
        address:
          type: string
        notes:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        size:
          type: string
        website:
          type: string
        address:
          type: string
        notes:
          type: string
...
    Deal:
      type: object
      required: [id, title, stage, value, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        contactId:
          type: ["integer", "null"]
        contactName:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        stage:
          type: string
          enum: [discovery, proposal, negotiation, contract, closed_won, closed_lost]
        value:
          type: number
        probability:
          type: ["number", "null"]
        expectedCloseDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    DealInput:
      type: object
      required: [title, stage, value]
      properties:
        title:
          type: string
        contactId:
          type: integer
        companyId:
          type: integer
        stage:
          type: string
          enum: [discovery, proposal, negotiation, contract, closed_won, closed_lost]
        value:
          type: number
        probability:
          type: number
        expectedCloseDate:
          type: string
        notes:
          type: string
    DealUpdate:
      type: object
      properties:
        title:
          type: string
        contactId:
          type: integer
        companyId:
          type: integer
        stage:
          type: string
          enum: [discovery, proposal, negotiation, contract, closed_won, closed_lost]
        value:
          type: number
        probability:
          type: number
        expectedCloseDate:
          type: string
        notes:
          type: string
...
    Invoice:
      type: object
      required: [id, number, contactId, status, total, createdAt]
      properties:
        id:
          type: integer
        number:
          type: string
        contactId:
          type: integer
        contactName:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceItem:
      type: object
      required: [description, quantity, unitPrice, total]
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
    InvoiceInput:
      type: object
      required: [contactId, status, items]
      properties:
        contactId:
          type: integer
        companyId:
          type: integer
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        tax:
          type: number
        dueDate:
          type: string
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        contactId:
          type: integer
        companyId:
          type: integer
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        tax:
          type: number
        dueDate:
          type: string
        notes:
          type: string
...